Conflicts:
	patientregistrationsystem/qdc/quiz/abc_search_engine.py

+++ patientregistrationsystem/qdc/experiment/views.py
@@ -5,6 +5,7 @@ from django.contrib import messages
 from django.http import HttpResponseRedirect
 from django.core.urlresolvers import reverse
 from django.shortcuts import render_to_response
+from django.db.models.deletion import ProtectedError
 
 from experiment.models import Experiment, QuestionnaireConfiguration, Subject, TimeUnit, QuestionnaireResponse
 from experiment.forms import ExperimentForm, QuestionnaireConfigurationForm, QuestionnaireResponseForm
@@ -53,7 +54,7 @@ def experiment_create(request, template_name="experiment/experiment_register.htm
 
 
 @login_required
-@permission_required('experiment.change_experiment')
+@permission_required('experiment.view_experiment')
 def experiment_update(request, experiment_id, template_name="experiment/experiment_register.html"):
     experiment = get_object_or_404(Experiment, pk=experiment_id)
     questionnaires_list = []
@@ -78,13 +79,20 @@ def experiment_update(request, experiment_id, template_name="experiment/experime
             if request.POST['action'] == "save":
 
                 if experiment_form.is_valid():
-                    experiment_form.save()
+                    if experiment_form.has_changed():
+                        experiment_form.save()
+
                     redirect_url = reverse("experiment_edit", args=(experiment_id,))
                     return HttpResponseRedirect(redirect_url)
 
             else:
                 if request.POST['action'] == "remove":
-                    experiment.delete()
+                    try:
+                        experiment.delete()
+                    except ProtectedError:
+                        messages.error(request, "Não foi possível excluir o experimento, pois há questões associadas")
+                        redirect_url = reverse("experiment_edit", args=(experiment.id,))
+                        return HttpResponseRedirect(redirect_url)
                     return redirect('experiment_list')
 
     context = {
@@ -182,7 +190,13 @@ def questionnaire_update(request, questionnaire_configuration_id,
                 return HttpResponseRedirect(redirect_url)
         else:
             if request.POST['action'] == "remove":
-                questionnaire_configuration.delete()
+                try:
+                    questionnaire_configuration.delete()
+                except ProtectedError:
+                    messages.error(request, "Não foi possível excluir o questionário, pois há respostas associadas")
+                    redirect_url = reverse("questionnaire_edit", args=(questionnaire_configuration_id,))
+                    return HttpResponseRedirect(redirect_url)
+
                 redirect_url = reverse("experiment_edit", args=(experiment.id,))
                 return HttpResponseRedirect(redirect_url)
 
@@ -222,13 +236,15 @@ def subjects(request, experiment_id, template_name="experiment/subjects.html"):
 
             if subject_responses:
                 if (questionnaire_configuration.number_of_fills is None and subject_responses.count() > 0) or \
-                        (questionnaire_configuration.number_of_fills is not None and questionnaire_configuration.number_of_fills == subject_responses.count()):
+                        (questionnaire_configuration.number_of_fills is not None and
+                            questionnaire_configuration.number_of_fills == subject_responses.count()):
 
                     number_of_questionnaires_completed = 0
 
                     for subject_response in subject_responses:
 
-                        response_result = surveys.get_participant_properties(questionnaire_configuration.lime_survey_id, subject_response.token_id, "completed")
+                        response_result = surveys.get_participant_properties(questionnaire_configuration.lime_survey_id,
+                                                                             subject_response.token_id, "completed")
 
                         if response_result == "N":
                             break
@@ -237,7 +253,8 @@ def subjects(request, experiment_id, template_name="experiment/subjects.html"):
 
                     if (questionnaire_configuration.number_of_fills is None and
                             number_of_questionnaires_completed > 0) or \
-                            (questionnaire_configuration.number_of_fills is not None and number_of_questionnaires_completed >= questionnaire_configuration.number_of_fills):
+                            (questionnaire_configuration.number_of_fills is not None and
+                                number_of_questionnaires_completed >= questionnaire_configuration.number_of_fills):
 
                         number_of_questionnaires_filled += 1
 
@@ -258,13 +275,10 @@ def subjects(request, experiment_id, template_name="experiment/subjects.html"):
     return render(request, template_name, context)
 
 
-@login_required
-@permission_required('experiment.add_questionnaireresponse')
 def subject_questionnaire_response_start_fill_questionnaire(request, subject_id, questionnaire_id):
     questionnaire_response_form = QuestionnaireResponseForm(request.POST)
 
     if questionnaire_response_form.is_valid():
-        date = request.POST['date']
 
         questionnaire_response = questionnaire_response_form.save(commit=False)
 
@@ -301,12 +315,6 @@ def subject_questionnaire_response_start_fill_questionnaire(request, subject_id,
         questionnaire_response.questionnaire_responsible = request.user
         questionnaire_response.save()
 
-        # Montagem da URL para redirecionar ao Lime Survey
-        date = date.replace('/', '-')
-
-        # redirect_url = '%s/index.php/survey/index/sid/%s/token/%s/lang/pt-BR/idavaliador/%s/datdataaquisicao/%s/idparticipante/%s' % (
-        #     settings.LIMESURVEY['URL'], questionnaire_config.lime_survey_id, result['token'], request.user.id, date, subject.id)
-
         redirect_url = get_limesurvey_response_url(questionnaire_response)
 
         return redirect_url
@@ -322,13 +330,23 @@ def get_limesurvey_response_url(questionnaire_response):
         questionnaire_response.token_id, "token")
     questionnaire_lime_survey.release_session_key()
 
-    redirect_url = '%s/index.php/survey/index/sid/%s/token/%s/lang/pt-BR/idavaliador/%s/datdataaquisicao/%s/idparticipante/%s' % (
+    redirect_url = \
+        '%s/index.php/%s/token/%s/idavaliador/%s/datdataaquisicao/%s/idparticipante/%s/newtest/Y' % (
         settings.LIMESURVEY['URL'],
         questionnaire_response.questionnaire_configuration.lime_survey_id,
         token,
-        questionnaire_response.questionnaire_responsible.id,
+        str(questionnaire_response.questionnaire_responsible.id),
         questionnaire_response.date.strftime('%d-%m-%Y'),
-        questionnaire_response.subject.id)
+        str(questionnaire_response.subject.id))
+
+    # redirect_url = \
+    #     '%s/index.php/survey/index/sid/%s/token/%s/lang/pt-BR/idavaliador/%s/datdataaquisicao/%s/idparticipante/%s' % (
+    #     settings.LIMESURVEY['URL'],
+    #     questionnaire_response.questionnaire_configuration.lime_survey_id,
+    #     token,
+    #     questionnaire_response.questionnaire_responsible.id,
+    #     questionnaire_response.date.strftime('%d-%m-%Y'),
+    #     questionnaire_response.subject.id)
 
     return redirect_url
 
@@ -337,7 +355,6 @@ def get_limesurvey_response_url(questionnaire_response):
 @permission_required('experiment.add_questionnaireresponse')
 def subject_questionnaire_response_create(request, experiment_id, subject_id, questionnaire_id,
                                           template_name="experiment/subject_questionnaire_response_form.html"):
-    experiment = get_object_or_404(Experiment, id=experiment_id)
 
     questionnaire_config = get_object_or_404(QuestionnaireConfiguration, id=questionnaire_id)
 
@@ -386,7 +403,7 @@ def subject_questionnaire_response_create(request, experiment_id, subject_id, qu
 @login_required
 @permission_required('experiment.change_questionnaireresponse')
 def questionnaire_response_update(request, questionnaire_response_id,
-                                template_name="experiment/subject_questionnaire_response_form.html"):
+                                  template_name="experiment/subject_questionnaire_response_form.html"):
 
     questionnaire_response = get_object_or_404(QuestionnaireResponse, id=questionnaire_response_id)
 
@@ -396,19 +413,21 @@ def questionnaire_response_update(request, questionnaire_response_id,
     survey_title = surveys.get_survey_title(questionnaire_configuration.lime_survey_id)
     survey_active = surveys.get_survey_properties(questionnaire_configuration.lime_survey_id, 'active')
     survey_admin = surveys.get_survey_properties(questionnaire_configuration.lime_survey_id, 'admin')
-    survey_completed = (surveys.get_participant_properties(questionnaire_configuration.lime_survey_id, questionnaire_response.token_id, "completed") != "N")
+    survey_completed = (surveys.get_participant_properties(questionnaire_configuration.lime_survey_id,
+                                                           questionnaire_response.token_id,
+                                                           "completed") != "N")
     surveys.release_session_key()
 
     questionnaire_responsible = questionnaire_response.questionnaire_responsible
     subject = questionnaire_response.subject
 
+    questionnaire_response_form = QuestionnaireResponseForm(None, instance=questionnaire_response)
+
     if request.method == "GET":
-        questionnaire_response_form = QuestionnaireResponseForm(request.POST or None, instance=questionnaire_response)
         fail = None
         redirect_url = None
 
     if request.method == "POST":
-        questionnaire_response_form = QuestionnaireResponseForm(request.POST, instance=questionnaire_response)
 
         if request.POST['action'] == "save":
 
@@ -422,11 +441,20 @@ def questionnaire_response_update(request, questionnaire_response_id,
 
         else:
             if request.POST['action'] == "remove":
-                questionnaire_response.delete()
+                surveys = Questionnaires()
+                result = surveys.delete_participant(
+                    questionnaire_configuration.lime_survey_id,
+                    questionnaire_response.token_id)
+                surveys.release_session_key()
+                result = result[str(questionnaire_response.token_id)]
+                if result == 'Deleted' or result == 'Invalid token ID':
+                    questionnaire_response.delete()
+                    messages.success(request, 'Preenchimento removido com sucesso')
+                else:
+                    messages.error(request, "Erro ao deletar o preenchimento")
                 redirect_url = reverse("subject_questionnaire",
                                        args=(questionnaire_configuration.experiment.id, subject.id,))
                 return HttpResponseRedirect(redirect_url)
-
 
     context = {
         "FAIL": fail,
@@ -550,16 +578,6 @@ def subject_questionnaire_view(request, experiment_id, subject_id,
     experiment = get_object_or_404(Experiment, id=experiment_id)
     subject = get_object_or_404(Subject, id=subject_id)
 
-    if request.method == "POST":
-
-        if request.POST['action'] == "remove":
-
-            experiment.subjects.remove(subject)
-
-            messages.info(request, 'Participante removido do experimento.')
-            redirect_url = reverse("subjects", args=(experiment_id,))
-            return HttpResponseRedirect(redirect_url)
-
     questionnaires_configuration_list = QuestionnaireConfiguration.objects.filter(experiment=experiment)
 
     subject_questionnaires = []
@@ -579,7 +597,9 @@ def subject_questionnaire_view(request, experiment_id, subject_id,
             can_remove = False
 
         for questionnaire_response in questionnaire_responses:
-            response_result = surveys.get_participant_properties(questionnaire_configuration.lime_survey_id, questionnaire_response.token_id, "completed")
+            response_result = surveys.get_participant_properties(questionnaire_configuration.lime_survey_id,
+                                                                 questionnaire_response.token_id,
+                                                                 "completed")
             questionnaire_responses_with_status.append(
                 {'questionnaire_response': questionnaire_response,
                  'completed': response_result != "N"}
@@ -591,11 +611,24 @@ def subject_questionnaire_view(request, experiment_id, subject_id,
              'questionnaire_responses': questionnaire_responses_with_status}
         )
 
+    if request.method == "POST":
+
+        if request.POST['action'] == "remove":
+            if can_remove:
+                experiment.subjects.remove(subject)
+
+                messages.info(request, 'Participante removido do experimento.')
+                redirect_url = reverse("subjects", args=(experiment_id,))
+                return HttpResponseRedirect(redirect_url)
+            else:
+                messages.error(request, "Não foi possível excluir o paciente, pois há respostas associadas")
+                redirect_url = reverse("subject_questionnaire", args=(experiment_id, subject_id,))
+                return HttpResponseRedirect(redirect_url)
+
     context = {
         'subject': subject,
         'experiment': experiment,
-        'subject_questionnaires': subject_questionnaires,
-        'can_remove': can_remove
+        'subject_questionnaires': subject_questionnaires
     }
 
     surveys.release_session_key()
